gcc warnings 64bit fix for Waypoint :: GetHash()

+++ Omnibot/Common/PathPlannerFloodFill.cpp
@@ -267,11 +267,11 @@ private:
 	//typedef FSBAllocator< std::pair< int, PlanNode* > > HashMapAllocator;
 
 #ifdef WIN32
-	typedef stdext::hash_compare<int> HashMapCompare;
-	typedef stdext::hash_map<int, PlanNode*, HashMapCompare/*, HashMapAllocator*/ > NavHashMap;
+	typedef stdext::hash_compare<uintptr_t> HashMapCompare;
+	typedef stdext::hash_map<uintptr_t, PlanNode*, HashMapCompare/*, HashMapAllocator*/ > NavHashMap;
 #else
-	typedef stdext::hash<int> HashMapCompare;
-	typedef stdext::hash_map<int, PlanNode*, HashMapCompare, stdext::equal_to<int>/*, HashMapAllocator*/ > NavHashMap;
+	typedef stdext::hash<uintptr_t> HashMapCompare;
+	typedef stdext::hash_map<uintptr_t, PlanNode*, HashMapCompare, stdext::equal_to<int>/*, HashMapAllocator*/ > NavHashMap;
 #endif
 	NodeList			m_OpenList;
 	NodeList			m_Solution;
